// ClientApp/src/app/core/nswag/plims-generated.ts
export interface UserShortInfoViewModel {
    id?: number;
    fullName?: string | null;
}
// ...
export interface ChemicalAnalysisViewModel {
    id?: number | null;
    regCode?: string | null;
    samplingDate?: Date | null;
    deliveringDate?: Date | null;
    analysisDate?: Date | null;
    wmrContent?: number;
    condensateContent?: number;
    massFractionOfMethanol?: number | null;
    massFractionOfMethanolError?: number | null;
    massFractionOfMethanolNotAvailable?: boolean;
    massFractionOfWater?: number | null;
    massFractionOfWaterError?: number | null;
    massFractionOfWaterNotAvailable?: boolean;
    massConcentrationOfCorrosionInhibitor?: number | null;
    massConcentrationOfCorrosionInhibitorNotAvailable?: boolean;
    place?: SamplingPlaceViewModel | null;
    point?: SamplingPointViewModel | null;
    user?: UserShortInfoViewModel | null;
}

export interface SamplingPlaceViewModel {
    id?: number | null;
    name?: string | null;
}

export interface SamplingPointViewModel {
    id?: number;
    name?: string | null;
}

export interface TreeItem {
    id?: number;
    name?: string | null;
    parentId?: number | null;
}

export interface ChemicalAnalysisTreeItem extends TreeItem {
    type?: ChemicalAnalysisTreeItemType;
}

export enum ChemicalAnalysisTreeItemType {
    Place = 0,
    Point = 1,
}